Add files via upload
Player One:
  arrow keys to move;
  'm' to shoot
  hold 'n' to reload
Player Two
  'w' 'a' 's' 'd' to move
  'q' to shoot
  hold '1' to reload 
The goal of the game is to eliminate your opponent as quickly as possible without dying yourself. This game is still under much construction but is being turned in in this state due to lack of time. I am sure I have made many sloppy mistakes but I feel like I could not find them, feel free to edit.